WIP: disable functional tests on travis move TLS checker class to bin script
